20230313_0026
int32 에서 uint8로 변경, 게임모드 상태 Enum추가, 플레이어 리스폰 (작업중)

+++ Source/Lakaya/Public/IndividualItem.h
@@ -35,11 +35,11 @@ private:
 
 public:
 	UPROPERTY(EditAnywhere, Category = "Item")
-	int32 ItemNumber;
+	uint8 ItemNumber;
 	
 public:
 	UFUNCTION(Category = "Item")
-	int32 GetItemNumber() { return ItemNumber; }
+	uint8 GetItemNumber() { return ItemNumber; }
 
 	UFUNCTION(Category = "Item")
 	void GetItem();
